Reduces function arguments
