+++ src/index.ts
@@ -1,8 +1,7 @@
 import createMatch from "./game/createMatch";
 import { logPlayers } from "./scoreboard/logs";
-import playMatch from "./playMatch";
+import playMatch from "./game/playMatch";
 import { getRandomPlayers } from "./game/util";
-
 
 // Entry point of the app. Pick two random players, create a match and play.
 const playRandomMatch = () => {
